Refactored junit and html reporters into html and csv reporters.
